my_telegram_bot/requirements.txt
python-telegram-bot~=20.8